New design: Humanizer members moved to instance
Adapted tests.
There is an issue with the rounding of the last unit: currently the code
only checks if the last unit in the count has been reached, and not if
it is the last unit we want (UnitToDisplay), so one of the test breaks.
That's why we need way more unit tests :-)